Fix build always expecting FW_DIR.
Date: Sun, 14 Jun 2009 11:51:33 +0100
From: Mike Auty <mike.auty@gmail.com>
Subject: [patch libraw1394-2] src/makefile.am expects FW_DIR to be set,
but configure only sets it if given --with-fw-dir

Here's a very small patch for the configure system of
libraw1394-2.0.{0,1,2}.  At the moment, if configure is called without
--with-fw-dir, then FW_DIR doesn't get specified.  The Makefile includes
the line INCLUDES=-I$(FW_DIR) and so in the compilation we get a -I not
followed by anything sensible.  That can cause compilation issues in
certain circumstances (see Gentoo bug 272540), so this patch ensures
that INCLUDES is only set if --with-fw-dir was specified.

Please let me know if there's any problems with the patch or if I've
submitted it to the wrong place or in the wrong way.  Thanks...

Mike  5:)

[1] http://bugs.gentoo.org/272540

Signed-off-by: Dan Dennedy <dan@dennedy.org>