Fix "make doc".
Reported by Guus Sliepen:  "make doc" failed due to missing doctype,
unknown elements, and duplicate element IDs in libraw1394.sgml.

The fix is to declare a recent DTD (matching the one which is used
in current Linux kernel documentation docbooks) and to make the
conflicting element IDs unique.

The latter part of the fix is just temporary.  In order to avoid the
conflict when the documentation is updated the next time, also fix the
kerneldoc comments of the respective API elements:  These are typedefs,
hence kernel-doc needs their comments prepended by "typedef ".

Tested with Gentoo's docbook-xml-dtd 4.5, docbook-xsl-stylesheets
1.75.2, docbook-sgml-utils 0.6.14, and openjade 1.3.2-r1.

Signed-off-by: Stefan Richter <stefanr@s5r6.in-berlin.de>
Signed-off-by: Dan Dennedy <dan@dennedy.org>