Include local firewire-*.h instead of system-wide <linux/firewire-*.h>
This guarantees that all features of libraw1394 are actually built in.
Before, some features and fixes would be silently dropped if too old
system headers (typically provided by a package called linux-headers)
were present.

An alternative would be to keep using system headers but add warnings
during the ./configure stage if old headers were encountered.  But this
helps only the person who builds libraw1394 (if there is a person
involved at all), not the users who have no reliable way to determine
how the library binary was built.

Another alternative would be to change the former soft dependency on
certain linux-headers versions into a hard dependency, i.e. fail the
build in absence of too old headers.  This would add an inconvenience
in setting up the build environment though:  The system headers would
have to be updated or a private copy of linux/firewire-*.h be specified
by way of the --with-fw-dir configure switch.

Anyhow.  The libraw1394 sources now already bring a suitable copy of
the two header files.  The --with-fw-dir configure switch is no longer
useful and is removed.

Signed-off-by: Stefan Richter <stefanr@s5r6.in-berlin.de>