Add 1394b speed codes to <libraw1394/{ieee,raw}1394.h>
This catapults the libraw1394 API into the year 2002.

Actually, passing speed codes of 3...5 into the relevant libraw1394
functions should be working already since the kernel gained 1394b
support a long time ago and libraw1394 does not check values.
The added definitions are only for clarity and to fully match the
argument type in the function declarations.

Signed-off-by: Stefan Richter <stefanr@s5r6.in-berlin.de>