Save and restore errno in raw1394_new_handle{,_on_port} for legacy applications
Since dual-stack capability was added to libraw1394, raw1394_new_handle()
and raw1394_new_handle_on_port() began to alter errno even when succeeding.
This breaks old application code which contains the bug of checking for
failure in errno rather than in the return code of said functions, or
similar bugs with wrong assumptions about errno.

While those applications should be fixed, it may not always be possible or
feasible to do so.  Hence add a workaround to libraw1394 which saves and
restores errno in said two functions.

From a superficial review of dispatch.c, it seems that these two functions
are the only ones where such a workaround may be needed.  However, this may
not be true if any fw_XYZ() function implementation differs from the
counterpart ieee1394_XYZ() function in the way that the former alters errno
during successful execution while the latter does not.  To be clear,
altering errno in absence of failure is absolutely allowed in library code
(except for signal handlers), yet it may be unexpected and be perceived as
a library or kernel regression if the application client code is buggy in
this regard.

Reported-by: Vladimir Romanov <blueboar2@gmail.com>
Signed-off-by: Stefan Richter <stefanr@s5r6.in-berlin.de>