Use <stdint.h> types instead of non-standard types
The __uint32_t type is not a standard type, and not necessarily
provided by all C libraries. However, <stdint.h> provides a set of
standard types, which are guaranteed to be available with all C
libraries.

This patch therefore changes __uint32_t to uint32_t, which allows to
fix the compilation with the Musl C library.

This patch is already used by the Alpine Linux distribution and the
Buildroot build system.

Original patch from
http://git.alpinelinux.org/cgit/aports/plain/main/libraw1394/fix-types.patch.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Stefan Richter <stefanr@s5r6.in-berlin.de>